+++ src/compiler/llvm_codegen_builtins.c
@@ -989,6 +989,9 @@ void llvm_emit_builtin_call(GenContext *c, BEValue *result_value, Expr *expr)
 		case BUILTIN_GET_ROUNDING_MODE:
 			llvm_value_set(result_value, llvm_emit_call_intrinsic(c, intrinsic_id.get_rounding, NULL, 0, NULL, 0), expr->type);
 			return;
+		case BUILTIN_IS_FPCLASS:
+			llvm_emit_simple_builtin(c, result_value, expr, intrinsic_id.is_fpclass);
+			return;
 		case BUILTIN_LOG:
 			llvm_emit_simple_builtin(c, result_value, expr, intrinsic_id.log);
 			return;